pyproject.toml
[project]
name = "vault-tools"
version = "1.1.0"
description = "CLI tools for working with Veeva Vault"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "prettytable",
    "python-dotenv",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "vapil @ git+https://github.com/veeva/vault-api-library-python.git",
    "vault-direct-data-accelerators @ git+https://github.com/veeva/Vault-Direct-Data-API-Accelerators.git",
]

[project.scripts]
vault-log-analyzer = "vault_tools.log_analyzer:main"
vault-ddapi = "vault_tools.ddapi_local.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vault_tools"]
